requirements.txt
fastapi==0.115.6
jinja2==3.1.4
pandas==2.0.3
pandas-gbq==0.23.1
python-dotenv==1.0.1
python-multipart==0.0.19
langchain==0.3.9
langchain_google_community==2.0.3
langchain-google-vertexai==2.0.8
langchain_openai==0.2.11
tqdm==4.66.5
uvicorn==0.32.1